+++ packages/core/src/types/index.ts
@@ -92,6 +92,7 @@ export type {
 export type {
   CollectionClientConfig,
   CollectionClientContentConfig,
+  CollectionStateClientConfig,
   ContextOf,
   DefinedResource,
   MessageLike,